# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "User Activity Tracking Go",
        "contact": {}
    },
    "paths": {
        "/api/client/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyJWTAuth": []
                    }
                ],
                "description": "Get client detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Get client detail",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hidayathamir_user-activity-tracking-go_internal_delivery_http_response.WebResponse-github_com_Hidayathamir_user-activity-tracking-go_internal_model_ResGetClientDetail"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "Login a client with name and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Login a client",
                "parameters": [
                    {
                        "description": "Client Login Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Hidayathamir_user-activity-tracking-go_internal_model.ReqLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hidayathamir_user-activity-tracking-go_internal_delivery_http_response.WebResponse-github_com_Hidayathamir_user-activity-tracking-go_internal_model_ResLogin"
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "description": "Register a new client with name, email, and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Register a new client",
                "parameters": [
                    {
                        "description": "Client Registration Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Hidayathamir_user-activity-tracking-go_internal_model.ReqRegister"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hidayathamir_user-activity-tracking-go_internal_delivery_http_response.WebResponse-github_com_Hidayathamir_user-activity-tracking-go_internal_model_ResRegister"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Hidayathamir_user-activity-tracking-go_internal_delivery_http_response.PageMetadata": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total_item": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "github_com_Hidayathamir_user-activity-tracking-go_internal_delivery_http_response.WebResponse-github_com_Hidayathamir_user-activity-tracking-go_internal_model_ResGetClientDetail": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_Hidayathamir_user-activity-tracking-go_internal_model.ResGetClientDetail"
                },
                "error_detail": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "error_message": {
                    "type": "string"
                },
                "paging": {
                    "$ref": "#/definitions/github_com_Hidayathamir_user-activity-tracking-go_internal_delivery_http_response.PageMetadata"
                }
            }
        },
        "github_com_Hidayathamir_user-activity-tracking-go_internal_delivery_http_response.WebResponse-github_com_Hidayathamir_user-activity-tracking-go_internal_model_ResLogin": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_Hidayathamir_user-activity-tracking-go_internal_model.ResLogin"
                },
                "error_detail": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "error_message": {
                    "type": "string"
                },
                "paging": {
                    "$ref": "#/definitions/github_com_Hidayathamir_user-activity-tracking-go_internal_delivery_http_response.PageMetadata"
                }
            }
        },
        "github_com_Hidayathamir_user-activity-tracking-go_internal_delivery_http_response.WebResponse-github_com_Hidayathamir_user-activity-tracking-go_internal_model_ResRegister": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_Hidayathamir_user-activity-tracking-go_internal_model.ResRegister"
                },
                "error_detail": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "error_message": {
                    "type": "string"
                },
                "paging": {
                    "$ref": "#/definitions/github_com_Hidayathamir_user-activity-tracking-go_internal_delivery_http_response.PageMetadata"
                }
            }
        },
        "github_com_Hidayathamir_user-activity-tracking-go_internal_model.ReqLogin": {
            "type": "object",
            "required": [
                "name",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_Hidayathamir_user-activity-tracking-go_internal_model.ReqRegister": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_Hidayathamir_user-activity-tracking-go_internal_model.ResGetClientDetail": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_Hidayathamir_user-activity-tracking-go_internal_model.ResLogin": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_Hidayathamir_user-activity-tracking-go_internal_model.ResRegister": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyJWTAuth": {
            "description": "JWT authorization",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}